Migrate core function to seperate project
