+++ Tomoyo.Core/Models/PhotographerProfile.cs
@@ -1,19 +1,19 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Tomoyo.Data;
+namespace Tomoyo.Core.Models;
 
 /// <summary>
-/// Base profile for application users, every user will have this object by default
+/// Cosplayer side profile bind with the user, created when user want to create
 /// </summary>
-public class BaseProfile
+public class PhotographerProfile
 {
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Id { get; set; }
-
-    public string UserId { get; set; }
-    public TomoyoUser User { get; set; } = null!;
+    
+    public string? UserId { get; set; }
+    public TomoyoUser? User { get; set; }
     
     [MaxLength(500)]
     public string? Bio { get; set; }
