+++ Tomoyo/Components/Account/IdentityNoOpEmailSender.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI.Services;
-using Tomoyo.Data;
+using Tomoyo.Core.Models;
 
 namespace Tomoyo.Components.Account;
