+++ Tomoyo/Components/Account/Pages/ConfirmEmailChange.razor
@@ -3,7 +3,7 @@
 @using System.Text
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject SignInManager<TomoyoUser> SignInManager
