+++ Tomoyo/Components/Account/Pages/ExternalLogin.razor
@@ -6,7 +6,7 @@
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject SignInManager<TomoyoUser> SignInManager
 @inject UserManager<TomoyoUser> UserManager
