+++ Tomoyo/Components/Account/Pages/Login.razor
@@ -3,7 +3,7 @@
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject SignInManager<TomoyoUser> SignInManager
 @inject ILogger<Login> Logger
