+++ Tomoyo/Components/Account/Pages/LoginWithRecoveryCode.razor
@@ -2,7 +2,7 @@
 
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject SignInManager<TomoyoUser> SignInManager
 @inject UserManager<TomoyoUser> UserManager
