+++ Tomoyo/Components/Account/Pages/Manage/DeletePersonalData.razor
@@ -2,7 +2,7 @@
 
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject SignInManager<TomoyoUser> SignInManager
