+++ Tomoyo/Components/Account/Pages/Manage/Disable2fa.razor
@@ -1,7 +1,7 @@
 ﻿@page "/Account/Manage/Disable2fa"
 
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject IdentityUserAccessor UserAccessor
