+++ Tomoyo/Components/Account/Pages/Manage/EnableAuthenticator.razor
@@ -5,7 +5,7 @@
 @using System.Text
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject IdentityUserAccessor UserAccessor
