+++ Tomoyo/Components/Account/Pages/Manage/ExternalLogins.razor
@@ -2,7 +2,7 @@
 
 @using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject SignInManager<TomoyoUser> SignInManager
