+++ Tomoyo/Components/Account/Pages/Manage/GenerateRecoveryCodes.razor
@@ -1,7 +1,7 @@
 ﻿@page "/Account/Manage/GenerateRecoveryCodes"
 
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject IdentityUserAccessor UserAccessor
