+++ Tomoyo/Components/Account/Pages/Manage/TwoFactorAuthentication.razor
@@ -2,7 +2,7 @@
 
 @using Microsoft.AspNetCore.Http.Features
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject SignInManager<TomoyoUser> SignInManager
