+++ Tomoyo/Components/Account/Pages/Register.razor
@@ -4,6 +4,7 @@
 @using System.Text
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
+@using Tomoyo.Core.Models
 @using Tomoyo.Data
 
 @inject UserManager<TomoyoUser> UserManager
