+++ Tomoyo/Components/Account/Pages/ResendEmailConfirmation.razor
@@ -5,7 +5,7 @@
 @using System.Text.Encodings.Web
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject UserManager<TomoyoUser> UserManager
 @inject IEmailSender<TomoyoUser> EmailSender
