+++ Tomoyo/Components/Account/Pages/ResetPassword.razor
@@ -4,7 +4,7 @@
 @using System.Text
 @using Microsoft.AspNetCore.Identity
 @using Microsoft.AspNetCore.WebUtilities
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject IdentityRedirectManager RedirectManager
 @inject UserManager<TomoyoUser> UserManager
