+++ Tomoyo/Components/Account/Shared/ExternalLoginPicker.razor
@@ -1,6 +1,6 @@
 ﻿@using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject SignInManager<TomoyoUser> SignInManager
 @inject IdentityRedirectManager RedirectManager
