+++ Tomoyo/Components/Account/Shared/ManageNavMenu.razor
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using Tomoyo.Data
+@using Tomoyo.Core.Models
 
 @inject SignInManager<TomoyoUser> SignInManager
