+++ Tomoyo/Tomoyo.csproj
@@ -9,12 +9,14 @@
 
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="9.*"/>
-        <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="9.*"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.*"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.*"/>
         <PackageReference Include="MudBlazor" Version="8.*"/>
         <PackageReference Include="Extensions.MudBlazor.StaticInput" Version="3.*"/>
+    </ItemGroup>
+    
+    <ItemGroup>
+        <ProjectReference Include="..\Tomoyo.Core\Tomoyo.Core.csproj"/>
     </ItemGroup>
 
 </Project>
